premier commit sur oli-novembre-2023

+++ app/Http/Controllers/Admin/RepliesController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Models\Reply;
 use App\Models\User;
 use Auth;
+use Carbon\Carbon;
 
 class RepliesController extends Controller
 {
@@ -17,7 +18,7 @@ class RepliesController extends Controller
      */
     public function index()
     {
-        //
+        
     }
 
     /**
@@ -38,10 +39,18 @@ class RepliesController extends Controller
      */
     public function store(Request $request)
     {   
-        Auth::user()->replies()->save(new Reply($request->only([
+        $reply = Auth::user()->replies()->save(new Reply($request->only([
             'body', 'conversation_id'
         ])));
-        return response()->json(['message' => 'Reply created successfully']);
+        $formattedCreatedAt = Carbon::parse($reply->created_at)->format('Y-m-d H:i:s');
+
+        return response()->json([
+            'message' => 'Reply created successfully',
+            'body' => $reply->body,
+            'name' => auth()->user()->name,
+            'created_at' => $formattedCreatedAt,
+            'id' => $reply->id,
+        ]);
     }
 
     /**
@@ -84,10 +93,20 @@ class RepliesController extends Controller
      * @param  \App\Models\Reply  $reply
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Reply $reply)
+    public function destroy(Request $request)
     {
+        $replyId = $request->input('reply_id');
+        
+        // Logic to find and delete the reply
+        $reply = Reply::find($replyId);
+        if (!$reply) {
+            return response()->json(['error' => 'Reply not found'], 404);
+        }
+    
+        // Perform the deletion
         $reply->delete();
-        return redirect('/forum');
-
+    
+        return response()->json(['message' => 'Reply deleted successfully'], 200);
     }
+    
 }
